refactor: moved and renamed GetJoinHelper class to EssentialsWebApiHelpers, added func for loaded assemblies
