+++ essentials-framework/Essentials Core/PepperDashEssentialsBase/PepperDash_Essentials_Core.csproj
@@ -199,7 +199,7 @@
     <Compile Include="Crestron IO\StatusSign\StatusSignController.cs" />
     <Compile Include="Web\RequestHandlers\AppDebugRequestHandler.cs" />
     <Compile Include="Web\RequestHandlers\GetFeedbacksForDeviceRequestHandler.cs" />
-    <Compile Include="Web\RequestHandlers\GetJoinMapHelpers.cs" />
+    <Compile Include="Web\EssentialsWebApiHelpers.cs" />
     <Compile Include="Web\RequestHandlers\GetTypesByFilterRequestHandler.cs" />
     <Compile Include="Web\RequestHandlers\GetJoinMapForDeviceKeyRequestHandler.cs" />
     <Compile Include="Web\RequestHandlers\DefaultRequestHandler.cs" />
